requirements.txt
discord
python-dotenv
pandas
psycopg2
sqlalchemy
bs4